add requesting csrf token for fixing login
The login process for humble bundle now requires a valid csrf token. You
just need to prove you are the client. This code simply pulls down the
login form and duplicates the csrf token given in that form within the
body of the login POST request that follows.

It just pulls out the token using a regular expression. The expression
should be robust enough to handle any slight changes and shouldn't break
as long as the key being used does not change.